Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    float x;
+    double x;
     printf("You entered: ");
-    scanf("%f", &x);
-    printf("%.4f", x);
-    return 0;
+    scanf("%lf", &x);
+    printf("%.10lf", x);
+    return 0 ;
 }
